+++ apps/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # walkeros
 
+## 0.3.5
+
+### Patch Changes
+
+- cli usage
+- Updated dependencies
+  - @walkeros/cli@0.3.5
+
 ## 0.3.4
 
 ### Patch Changes