+++ apps/demos/destination/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @walkeros/destination-demo
 
+## 0.3.5
+
+### Patch Changes
+
+- cli usage
+
 ## 0.3.4
 
 ### Patch Changes