+++ apps/demos/source/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @walkeros/source-demo
 
+## 0.3.4
+
+### Patch Changes
+
+- cli usage
+
 ## 0.3.3
 
 ### Patch Changes