+++ apps/quickstart/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @walkeros/quickstart
 
+## 0.3.3
+
+### Patch Changes
+
+- cli usage
+
 ## 0.3.2
 
 ### Patch Changes